VariationParser: extracted Variation Frequencies parsing into a new class
